backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pymysql
python-dotenv
alembic
argon2-cffi
razorpay
PyJWT
